INT: build default values for constants in ImplementMembersFix
